testing: move CorrelatedStack mock classes and data generation to tests.data
